convert src/rogue.pct from PICT to PNG and remove it from src dir
It contains the splash image shown in the Rogue version publixhed by Epyx.
However, this file is not used in build or run time. It was probably the master
file used to generate `src/rogue.pic`, which shipped with the game and is loaded
only at run time.

The file is a PICT image format used in early Apple Macintoshes.
See https://en.wikipedia.org/wiki/PICT

ImageMagick was used to identify and convert it to PNG.