query keyboard LED status via X11 C API
Executing xset on every tick is a rather expensive operation, as it has to
spawn a new process and open and close a TCP connection each time, slowing the
game down significantly.  Additionally, if $DISPLAY is set but no X server is
present, error messages appear continuously and make the game unplayable.

To resolve this, we can use the X11 C API to open an X display connection on
startup, and keep it open for the duration of the game.  Then the keyboard LED
status can be queried directly via XGetKeyboardControl().  With PuTTY, if X11
forwarding is enabled ($DISPLAY is set) but no X server is running on the
client, this still results in a single error message on startup, but it quickly
disappears after raise_curtain().