fix vorpalize scroll selecting invalid target monster
The Vorpalize scroll enchants a weapon to do more damage to a specific
type of monster.  This monster is randomly picked in pick_mons() from
the lvl_mons string.

Problem is, this string contains spaces, and sometimes (surprisingly
often) it lands on one.  Identification then causes an out-of-bounds
access on the monsters array.  In the original DOS version, this just
results in a garbled name, and a weapon with no special abilities.
On modern systems however, we get a segmentation fault.

This patch resolves the issue by introducing a new string with no
spaces, vorp_mons, from which targets are selected.  This string
contains the full alphabet, so that all monster types are considered
potential targets.

From disassembling earlier Rogue PC versions, specifically v1.0 by Jon
Lane, we can deduce that lvl_mons originally contained no spaces when
the Vorpalize scroll was introduced, so re-introducing this string with
no spaces is believed to reproduce the intended behavior.